Move authentication actions to its own actions-file
