+++ src/main/typescript/actions/authenticationActions.ts
@@ -1,0 +1,47 @@
+/*
+ * Copyright (C) 2018 DANS - Data Archiving and Networked Services (info@dans.knaw.nl)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { ReduxAction } from "../lib/redux"
+import { loginURL, userURL, logoutURL } from "../constants/apiConstants"
+import { AuthenticationConstants } from "../constants/authenticationConstants"
+import axios from "axios"
+
+export const authenticate: (userName: string, password: string) => ReduxAction<Promise<any>> = (userName, password) => ({
+    type: AuthenticationConstants.AUTH_LOGIN,
+    async payload() {
+        await new Promise(vs => {
+                setTimeout(vs, 1000)
+            },
+        )
+
+        // TODO temporary do a timeout to simulate server I/O
+        const { data } = await axios.post(loginURL, {}, { auth: { username: userName, password: password } })
+        return data
+    },
+})
+
+// TODO signout also requires a server call...
+export const signout: () => ReduxAction<Promise<void>> = () => ({
+    type: AuthenticationConstants.AUTH_LOGOUT,
+    async payload() {
+        // TODO temporary do a timeout to simulate server I/O
+        await new Promise(vs => {
+                setTimeout(vs, 1000)
+            },
+        )
+        await axios.post(logoutURL)
+    }
+})
+
