+++ src/main/typescript/components/Header.tsx
@@ -19,7 +19,8 @@ import { Link, NavLinkProps } from "react-router-dom"
 import { AppState } from "../model/AppState"
 import { connect } from "react-redux"
 import { Action } from "redux"
-import { getUser, signout } from "../actions/userActions"
+import { getUser } from "../actions/userActions"
+import { signout } from "../actions/authenticationActions"
 import "../../resources/css/header"
 import { ReduxAction } from "../lib/redux"
 import { depositOverviewRoute, homeRoute, loginRoute, registerRoute } from "../constants/clientRoutes"
