+++ src/main/typescript/components/login/LoginPage.tsx
@@ -17,7 +17,7 @@ import * as React from "react"
 import { Component } from "react"
 import { ReduxAction } from "../../lib/redux"
 import { connect } from "react-redux"
-import { authenticate } from "../../actions/userActions"
+import { authenticate } from "../../actions/authenticationActions"
 import { Redirect, RouteComponentProps } from "react-router"
 import { AppState } from "../../model/AppState"
 import { homeRoute } from "../../constants/clientRoutes"
